Fix glint and pyformat issues
